no await, no Later, no intermediate states
This differs between the VSCode-run-once result and command line result
in that the IDE debugger result will reliably give the intermediate
states. The command line operates in a much more async nature so it is
necessary to find a way to completely ignore any states besides the end
state that we are looking for.

Also worth noting that each test should get its own database since they
seem to be getting deleted in the middle of other tests. When the setup
routine involves deleting all of the databases and the tests are
starting at the same time as one another there is a risk of a db being
deleted in the middle of another test and causing it to fail.

Will need to keep an eye on these tests once they all start passing
locally to see if the same behavior occurs on the CI server and whether
or not the tests will pass 100% of the time or have more sporadic issues